Improve scope information for language plugins.
This adds an `icon()` method to `ScopeChainEntry`, and extends the
`ScopeChainSideparPane` with an optional `<img>` element that shows
this icon. This implements the recommended UX improvements to make
it easier to differentiate between the scopes themself and the items
listed in the scopes.

It is implemented via a new method `getScopeInfo()` method in the
language plugin extension API that returns the scope type, a user
visible type name, and the optional icon.

Before: https://imgur.com/a/xIvhdM4
After: https://imgur.com/a/6WHoceO
Bug: chromium:1136484
Change-Id: I5927e1502bc88b0d802aca5dcbc220df8d7c7f5d