Update Rollup config to allow for multiple entrypoints in a module
Modules can have multiple entrypoints. The current state of affairs
has `module-legacy.js` and `module.js` files in many folders. To
enable us to move `-legacy.js` files into the build graph, we need
to build them separately (this is currently being done by the
out-of-process `build_release_applications.py` script). However,
we don't want the `devtools_entrypoints` for a legacy file to also
include the contents of the module entrypoint.

Therefore, assume that imports of the entrypoint itself are considered
external and will not be bundled in. Additionally add some small
types to several variables to get in-editor typechecking.

R=aerotwist@chromium.org

Bug: 1131500
Change-Id: I053eae4d8b237148ce6c41eede9388b2a7e5914d