Typecheck extensions/LanguageExtensionEndpoint.js with TypeScript
This also includes a refactoring of the ExtensionAPI. In the file
itself, it was creating an "apiPrivate" that would indirectly
reference several enums. Instead of referencing these enums via
an untype-able object, we can declare them as separate enums.

This actually caught a bug, where one of the usages of Events
was looking up InspectedURLChanged, which was non-existent.

R=pfaffe@chromium.org,szuend@chromium.org

Bug: 1011811
Change-Id: I775e71898bff67cbb0a15ed823bc0d3d10e78510