Remove backtick for lock terms in en-US.ctc.json files
In Lighthouse project, backtick is being used to indicate something that should not be localized.
(See 'inline code block' section in https://github.com/GoogleChrome/lighthouse/tree/master/lighthouse-core/lib/i18n)

We can adopt this with minor modification. When the lighthouse report is being rendered, the backtick in message will be converted to a special html tag and used to style the code snippets. Since we don't want to do special styling for lock terms, backtick can be removed after the term being converted to placeholder in en-US.ctc.json, then locale.json would have the term that is not translated and without the backtick .

For example
Before change:
After change:





Bug: 1136655
Change-Id: I68ea305962beec4115d0a50f24e34546b92cadf7