CSS: Fix inspector-stylesheet links on devtools reload
With the recent changes for supporting constructed stylesheets, the
"inspector-stylesheet" links created when adding new styles to a page
using the styles sidepane where being incorrectly marked as
"<style>" tags after closing and reopening the devtools. This
CL fixes the issue by moving the check for a valid rule location and
style sheet before the logic for constructed style sheet nodes.

Before:
After:
Change-Id: Iaa1f54d960dfa33588da78efe7b5e2e95179e175