Implement support for the JavaScript formatter API
For the formatter API, we introduce two new language plugin API methods
for querying the structure and the value of variables seperately. To
display a variable, the getTypeInfo method is first used to retrieve
structural information. The structure is used initially to display a
variable in the scope view, watch window, or console. If a node in the
structure tree cannot (or shouldn't be) further expanded, the second
method, getFormatter, is used to fetch a piece of JavaScript code from
the plugin that when passed to evaluateOnCallFrame produces a variable
value to be shown.

Bug: chromium:1127915, chromium:1137754
Change-Id: Id0a77ccdababe686e78003bcfac33200817dc8fa