Fix the lookup of debugger model and debugger id
The frontend maintains a map between debugger ids and
corresponding debugger models. However, sometimes the
debugger id changes in the backend, the debugger model remains
the same as the frontend and backend share the debugger model.
This means for the map to work, the frontend and backend
need to resync on debugger ids.

This fix uses existing CDP call (Debugger.enable) to resync
on debugger id. Future improvement may have a different CDP
call for this.

Repro exists when we enable the setAttachDebugHeader. Due to a
backend issue that yet to be fixed, the test associated with this
fix will be check in shortly after this fix and the backend fix
currently under review.

Bug: https://bugs.chromium.org/p/chromium/issues/detail?id=1137300
Change-Id: If88dd6b6c42421fe81b3d7fa9271a1d522dd27ce