Fix the lookup of debugger model and debugger id
The frontend maintains a map between debugger ids and
corresponding debugger models. However, in the backend,
sometimes, a new (debugger id, model) pair is made. The
frontend does not actively update the map. As a result,
while the frontend may still have a valid model, it couldn't
locate the model using a debugger id sent from the backend.

This is particular bad when we print the asynchronous stack
id in CallStackSidebarPane. We need to resync on debugger ids.

This fix uses existing CDP call (Debugger.enable) to resync
on debugger id. Future improvement may have a different CDP
call for this.

Bug: https://bugs.chromium.org/p/chromium/issues/detail?id=1137300
Change-Id: If88dd6b6c42421fe81b3d7fa9271a1d522dd27ce