Add clip-path property values for autocomplete menu
Enter values for the clip-path property into the manually maintained
list that is used to construct Styles editor autocomplete menus.
While the requestor asked for the list to include the same values
as appear in Firefox, this fix only adds the values that are
supported by Chromium.

Screenshot before fix:
https://storage.googleapis.com/issue_954126_screenshots/clip-path_menu_before_fix.png
Screenshot after fix:
https://storage.googleapis.com/issue_954126_screenshots/clip-path_menu_after_fix.png

Bug: 954126
Change-Id: I7581ee92452c7796114e50919503110babef366f