[wasm] Add toolbar hint for UISourceCodes loaded by language plugins.
In the right bottom corner of the text editor we display a hint for
source mapped files linking back to the original URL of the script
for which the source map was loaded, but so far we didn't have the
same functionality for language plugins. This adds a message and a
link to the script URL in the same fashion.

Bug: chromium:1136484
Screenshot: https://i.imgur.com/XJU7NH9.png
Change-Id: I57e66200d93db0458d77ceb32aec9f070d8f224a