Inspector overlay: include common.css in main.ts
After we include common.css in main.ts we are able to remove
common.css as a separate resource that gets shipped with Blink.
For now common.css is still a separate resource and with this CL
it will be included twice into the page. I will make backend changes
to remove common.css once this CL is rolled.

Change-Id: I0f099f27eee192dcbb8298e21bdfafdf00324c3a