Show icons explaining flexbox in Styles panel
This CL adds an icon for the flex-direction: row property that
shows up when editing the styles. The icon is automatically rotated
based on the element's direction. The feature is behind a new experiment
flag called cssFlexboxFeatures. Note that icon is not final.

Video: https://i.imgur.com/J17Zq1U.mp4

Bug: 1139945
Change-Id: I674215f62b2de6fd2f14884974eda186ad39ce3e