Grid: Simplify the label color logic to reduce the bundle size
The css grid overlay's label background color is a slightly lighter
shade than the main line color.
To achieve this, the overlay blends the line colors with a
semi-transparent white color.
This requires importing the Color.js module.

It turns out that because the 10 default colors we use are already
really light, making them slightly lighter makes almost no difference at
all.
So, in order to make the overlay bundle smaller, I'm removing this logic
altogether.

In the process, I'm also removing the color-contrast logic as we want
the label text color to always be dark.

Before/after comparisons: https://imgur.com/a/RChe3zt

Bug: 1139906
Change-Id: I5bf62f3a5df8e137dc02f4fcda3a1fa32661333c