Improvements for the Icon component
Wanted to suggest a few improvements to the Icon component:

- removed mentions of someIcon in the code.
- unified the source for styles and (subjective) made the data flow
more top-down.
- added support for changing the icon color via CSS as I have started
to use it and it is not always possible to change it via JS + it feels
like setting the color can be done without a rerender.

Follow-up: I am thinking about making the difference between masked
and background image components more explicit. WDYT?
Change-Id: I3b15dc3a1b1039a59fd5e2c5e8d2136e696f7989