Removed border-bottom from header, added border-top to container wrapper
In order to stop the contents of settings-container-wrapper from scrolling over its parent header div when minimum font size is set to huge, I removed the border-bottom from the header element and instead added the same border to the container-wrapper top.  It looks exactly the same and solves the issue.

Added my name to the AUTHORS file

R=mathias@chromium.org

Bug: 1133677
Change-Id: I79c99dd04baf34a9a6c9fe6f623cf7f4964965b1