[wasm] Hide gutter context menu options that don't work.
The following context menu options don't currently behave correctly with
WebAssembly or simply don't make sense, and are thus hidden from the UI
for now:

- 'Add condition breakpoint'
- 'Add logpoint'
- 'Edit breakpoint'
- 'Never pause here'
- 'Blackbox script'
- 'Continue to here'

Introduce a `DebuggerWorkspaceBinding.scripts()` method to iterate all
the scripts that are related to a given `UISourceCode`. We use that
method to check that every script related to the current `UISourceCode`
is JavaScript, and thus supports the full range of debugging
capabilities.

Bug: chromium:1136484
Change-Id: If6277b43b2fb2fb73e9f3aae53e09074b614f6ed
Fixed: chromium:1142275