Updated template for text buttons to allow for mousdown event listenet
Due to text buttons being created with an event listener that listens for
clicks the function that runs when the cancel button is clicked ran after
the function that adds highlighting to the input.  I added an optional 5th
parameter in the template for text buttons to allow for the option to change theevent listener from click to any other event listener.  Following this I updated the creation of the cancel button so that it is created with a mousedown event
listener allowing it to fire prior to being highlighted and now everything fires
in the correct order.


Bug: 1131388
Change-Id: I4428f1add8115735ab8ad2013dd6bedc67aa087b