Fix search index for JSON sources when no match found
Previously the search index was being incorrectly for
JSON sources with no match. It was being set to 0. This
resulted in search results showing "1 of 0". This was
fixed by setting the search index to -1 for no matches.

Before fix screenshot: https://imgur.com/a/yNzXNAV
After fix screenshot: https://imgur.com/a/Zc1ZyZq

Bug: 838311
Change-Id: I84bae4a639a001adc12eb13b2a96447562c60fba