Update Console Error/Warnings Counters CSS for consistency
This CL is a follow up to the minor refresh work on the main toolbar
Errors, Warnings, and Issues counters [1].

The CSS has been updated to prefer class selectors instead of element
selectors to match the style of other module CSS
(e.g. consoleView.css and elementsPanel.css prefer class specificity).
This will help avoid rule conflicts with common shared CSS and make it
easier for third party themes to override.

The errorWarningCounter.css file has been renamed to
warningErrorCounter.css to match the name of the module and follow file
naming conventions of other modules (e.g. ElementsPanel.js and
elementsPanel.css). It's unclear why there was a naming difference here
but it has been present for several years.

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2436646

Bug: 1133060
Change-Id: Ic08f47ecff2414dd4d3efb70fbce98e5e6bf6b82