Fix component bridges and use of legacy interfaces
The ColorSwatch.ts component uses the `Common.Color.Color` interface.
The bridges generator does not support deeply nested interfaces. So this
CL adds special casing for interfaces that we know we might have to deal
with from "legacy land" and makes sure they still get outputted
correctly.

Currently we only allow nested interfaces that start with `Common.`,
because I'd like to avoid their use in the new world if possible, but we
can easily expand this if required.

Once I got the type being compiled correctly, I then realised that we
also needed to add the imports into the outputted file, so the code now
checks for usage of Common, finds the matching import, and pulls it
over.

Finally, I had to update ColorSwatch.ts. It used public properties, which the
bridge generator doesn't support, so I swapped it to private properties with
getters. I don't love this change, but I think that's better rather than invest
more time in the (temporary) bridge generator code.

Note: this bug made it in because there's a bug in the bridges PRESUBMIT that
means if it errors it doesn't fail the PRESUBMIT. I have another CL incoming to
fix that, but I need to fix the actual component first so that when I fix the
PRESUBMIT I don't just block CQ for everyone!
Change-Id: I7376b0b7bee78bfe9d106fad6e2233a2b04fcf42