Fix bridges presubmit error not failing presubmit
If we get an unexpected error when re-generating bridges, we log it, but
don't make the actual process error, so the presubmit doesn't error!

This caused an issue to sneak in which https://crrev.com/c/2502042
fixes. Until that CL lands and this is rebased, this CL will fail on the
presubmit.

Change-Id: Ib1a3515ff8c93dfefcf3fa20dbba7c6d8ab9353b