adds type to remoteObject when storing as global variable
This change adds the type of the remote object being stored as a global
variable. Previously, if there were two objects able to be stored, the
context menu would contain duplicate selections leaving the user unable
to distinguish between the two. By adding the object type, the user can
determine what exactly they will be storing as a global variable.


Bug: 813374
Change-Id: Ie644e2b7b32fcc64d3c1cac562fb2398ff5fc71f