Clear network search input on new instance of DevTools
Move _onSearchInputClear() off of _resetSearch() and onto the clearButton eventListener because while the two methods are related, they are mutually exclusive.

Add _save() method to _onSearchInputClear() which saves an empty string as the query value to local storage.

This bug has to do with the search input in the Network section of DevTools. The search input is populated with a value which is stored in local storage as the query object. Each time DevTools loads, it requests the value from local storage and populates the search input with the value.  The bug occurs because the value returned from local storage is the value from the previous search. Pressing the clear button had no effect on the query value stored in local storage.

When the clear button is pressed, this patch saves the query value to local storage as an empty string in the query object.  Any new instances of Devtools, and any instances opened in NTP will find the search input of the Network panel empty, after the clear button is pressed.  Current instances or instances where the clear button is not pressed will retain the previous value and allow user access to searched results.

Bug: 1099160
Change-Id: I7afbd49fad7c68286d7854ab7b6c898fe3ad70ae