Migrate 3 web_tests to the corresponding unit and e2e tests in frontend
While working on refactoring the color and var swatches recently, I had
to disable a few web tests in chromium.src.
Instead of fixing them and re-enabling them now, I'm taking this
opportunity to convert them to the appropriate unit or e2e tests in the
frontend (this helps reduce the dependency for frontend-only features on
chromium.src).

Here is the list of disabled web_tests and what I did for each:
1. third_party/blink/web_tests/http/tests/devtools/elements/styles-1/color-swatch.js
   Converted to both new unit and e2e test cases for what wasn't already
   covered.
2. third_party/blink/web_tests/http/tests/devtools/elements/styles/url-color-swatch.js
   Added new e2e test cases for what wasn't already covered.
3. third_party/blink/web_tests/http/tests/devtools/elements/css-variables/color-swatch.js
   Also added new e2e test cases to cover this.
3. third_party/blink/web_tests/http/tests/devtools/sources/color-swatch-position.js
   This one is different as it tests editing CSS code in the Sources
   panel. Because the current e2e helpers lack support for this, I will
   fix and re-enable this web_test in chromium.src. It doesn't seem
   worth it right now to create a bunch of new helpers just for the
   color swatch. But we should do this too eventually.

Backend CL to come soon to delete/re-enable the tests.

Bug: 1137438
Change-Id: Ic260e460ba215dd9a317ae2843d927c2d1bace3d