Support page script in fetch handler
The change is two-fold:
1) enable setAttachDebugStack CDP call(We may move the CDP
method in the future as it makes sense for it to always be
enabled). An explainer below provides the context of this work.
2) update the debugger id to debugger model mapping in the
frontend (explained below).

The frontend maintains a map between debugger ids and
corresponding debugger models. However, in the backend,
sometimes, a new (debugger id, model) pair is made. The
frontend does not actively update the map. As a result,
while the frontend may still have a valid model, it couldn't
locate the model using a debugger id sent from the backend.

When we print the asynchronous stack id in CallStackSidebarPane.
We need to resync on debugger ids. We use a existing CDP call
(Debugger.enable) to resync on debugger id. Future improvement
may have a different CDP call for this.

Screenshot: https://imgur.com/oU5wETQ

Explainer:https://tinyurl.com/y3zpcfdz

Corresponding backend change:
https://chromium-review.googlesource.com/c/chromium/src/+/2465285

Bug: https://bugs.chromium.org/p/chromium/issues/detail?id=1137300
Change-Id: If88dd6b6c42421fe81b3d7fa9271a1d522dd27ce