Populate image preview toolbar content early
This reduces the flicker that we see when selecting an image on the
Network panel for preview. The reason for the flicker is the
asynchronicity between creating the toolbar and populating its content.

There is still a slight flicker left over that is caused by removing the
widget and then asynchronously later creating a new widget when changing
the selection. A lot more re-architecturing would be required to fix
this.

R=aerotwist@chromium.org

Bug: chromium:1142729
Change-Id: Idefd4731577d6eada8ab0a15e7be2d5ee924e46e