Support fake connections to the backend
Many unit tests are blocked by the fact that they require the SDK but
the SDK requires a live connection to the backend. This CL introduces a
fake connection that can be used instead, and instantiates an
SDK.CookieModel.CookieModel to demonstrate how it can be used.

While this CL has side-effects, due to other migrations they are far
less problematic. Specifically, creating a Root.Runtime.Runtime.instance
and a Common.Settings.Settings.instance is now overwritable thanks to
the forceNew flag (whereas previously it was not). This means that
anyone who wishes to reinstantiate these entities for a different test
can do so easily. Nonetheless there is clean-up code here for those
instances to remove them at the end of the test.

Change-Id: I6734c7f69f4f6b8e6433976aa5f38a3dd2e2f99a