Use TypeScript-authored files in issues/
Since we have changed `build_release_applications` to ignore the files
listed in the `modules` array (
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2461783)
the files listed in `modules` are only used for Closure. This means
that if we remove the files from `modules`, Closure will no longer
know about it.

There is one caveat: the runtime still does the contains-check for
the legacy file. Therefore, we have to include the `issues-legacy.js`
file in the `modules` array (until we have removed its extensions, see
crbug.com/1134103). To ignore Closure warnings, we also have to
add the legacy file to the `skip_compilations` array.

Note that we can only do this for modules which satisfy these
requirements:
1. They are referenced directly by the entrypoint JSON files
2. There are no modules that are still type checked by Closure
that depend on this module

R=aerotwist@chromium.org,szuend@chromium.org

Bug: 1011811, 1127902
Change-Id: I348f2735312fc6147915a88b1377e85c02161d8a