Prevent NetworkLog reset in NetworkPanel constructor
After https://crrev.com/c/2503513 the NetworkLog would be reset in the
NetworkPanel constructor. This prevented DevTools from displaying
requests which were recorded before opening the NetworkPanel for the
first time. This CL returns to the previous behaviour of not resetting
the NetworkLog in this scenario.

In addition, the hint displayed when the NetworkLog is empty, is
updated for the case when recording is turned off.

Fixed: chromium:1144720

Change-Id: I3ed4e8a89c1e6aa9bd139c44d32dafd84749e3b3