Add align-self icons to the Styles panel
This CL adds item-level icons for the align-self property.
The position of the icon is the same as the position of the
align-items icons for the parent node. Parent node's styles
are, therefore, used when a parent node is available.
Parent node's computed styles are cached as well.

Screenshot: https://i.imgur.com/OAD4eqS.png
Bug: 1139945
Change-Id: I83000ced65b92b75976e8733b54dfd134694b271