Update frame tree when frames are written to via JS
Even though writing to a frame's document via JS is not considered a
navigation by the backend, it can change the information DevTools is
showing about a frame in the frame tree. We are adding a new CDP event
'Page.documentClosed' in the backend, which in the frontend triggers
the same cascade a frame navigation would. This ensures that
ResourceTreeFrame and therefore the frame tree is kept up to date.

Backend CL: https://crrev.com/c/2498482

Bug: chromium:1140540, chromium:1140481

DISABLE_THIRD_PARTY_CHECK=protocol update

Change-Id: Idf75e9101508923f83abb2525116611b4e2ff7a9