Show icons for align-content:start|end
This CL adds icons for align-content:start|end values which are not
supported in Chromium yet. They are expected to work in the same way
as the flex-start|flex-end for align-content (tested in Firefox).
Also, it fixes a bug in align-content icon positioning for row-reverse
and column-reverse flex directions and adds a few tests.

Bug: 1139945
Change-Id: I34419fcc9fc9ab7a9ea506bbc7e6f3830f0c9497