Show API availability in FrameDetailsView
This CL shows the gated APIs/features for SAB, SAB+postMessage and
document.domain in the FrameDetailsView. The CL also introduces
a deprecation warning if SAB+postMessage is currently available,
but the context is not cross-origin isolated.

DISABLE_THIRD_PARTY_CHECK=required for dependency

Screenshots: https://imgur.com/a/7UJ2JpZ
Bug: chromium:1139899
Change-Id: Id4c892b9a9842cbf655d40debc315cd24b7cf978