Wrap scrubber in overflow wrapper on animation timeline
The animation timeline uses a scrubber element whose width provides the
dark overlay that goes ontop of the animation timeline as the scrubber
proceeds from left to right.

https://crbug.com/1140888 is a bug where this dark overlay extends
beyond the length of the animation timeline, such that if a user tabs
through to focus on an event that occurs towards the end of the
timeline, they will see a blank space at the end. This blank space is
created by the scrubber's overlay. (See the bug report for a video).

The fix is to wrap the scrubber in a container that's the exact size of
the timeline itself, and add `overflow: hidden` to it. That ensures that
as the scrubber moves right, only the overlay that fits onto the
timeline is visible, and the rest is hidden. This prevents it growing
beyond the dimensions of the timeline and thus the user is unable to get
to the state in the bug above where they see blank space.

Fixed: 1140888
Change-Id: Ie1d373e0be2db2ebc163e81042cee6fdbf320a02