Fix WebAudioModel
The TypeScriptification of /webaudio involved changing over some model
functions to using events. I mistakenly dispatched the event itself to
the listeners, whereas what I should have done was to obtain the data
from the event and dispatch *that* instead.

This CL, then, extracts the data from the events in the WebAudioModel
and dispatches it to listeners.

Bug: 1144522
Change-Id: I7f6df3975bb7d106734e2a6c823bb348416536f0