Fix event propagation for link from FrameDetails to Network request
Consume the event within the link from FrameDetails to Network request.
This prevents the event from bubbling up to the Network Panel's filter
box.

Also increases the width of the Network Panel's filter input field, so
that it fully fills its parent element. This ensures that clicking on
the right part of the filter input is handled correctly.

Fixed: chromium:1139813, chromium:1145007
Bug: chromium:1145525

Change-Id: Iec08b8b19a6bab7b711e6e20f620f1e918a40a33