[unit test] Isolate connection and environment helpers
Currently the mock connection also includes booting a minimal
environment. This is mostly fine, because if you require the mock
connection you also require the environment. However, it is also
plausible that someone will want to write unit tests that require an
environment (e.g. some settings, perhaps) but not a mock connection.

This CL, therefore, separates out the mock connection and the
environment initialization, such that someone can initialize them
separately if they so desire. It also provides a describe hook for the
environment, and also removes another helper that initialized the
settings alone.

R=jacktfranklin@chromium.org

Change-Id: I18de0aa48a92d18462eacfb84ac469481e848561