Fix resettting user focused cell in data-grid.
This CL started out as adding some tests to some areas that were lacking
test coverage, but then I realised we had a bug where if the user is
foucsed on a cell but that cell were to be removed or hidden, our focus
tracking would break. So this CL adds tests and code for that.

I purposefully kept the logic simple; if they go out of bounds, we put
them on the last row/column we can. Else, we reset them to [0, 0]. This
may prove too simplistic and need fixing, but we can defer that until
we're using the data-grid more and can try it out in DevTools properly.

Change-Id: I35f7fcaa3b4ed6f242d93063bb5a6c83f0abbe8d