[issues] Introduce substitutions in markdown issue descriptions
This CL enables dynamic replacements in markdown issue descriptions.
This can be used to insert information relating to a specific issue,
e.g. a violating URL or a offending CSP directive.

The code in this CL is very likely to be replaced with a l10n
placeholder mechanism, once a translation pipline for markdown
issue descriptions is in place.

R=sigurds@chromium.org

Bug: chromium:1125465
Change-Id: Id71369a1fe2b485fdeea21c8b39bb0d8de3295b7