Refactor CSSAngle for more modulatiry
This CL separates swatch and editor from the main CSSAngle, so that we
can better reuse them (mostly editor) in a compositied linear gradient
tool later.

This CL also adds more unit tests for keyboard accessibility.

After this CL, it will also become safer and easier to fix the unit
conversion bug , which will be addressed in a follow-up CL.

Bug: chromium:1126178
Change-Id: I2e1ce5880b944ed9bc248d63e0927903014f19aa